+++ src/plugins/components/PluginsListPage/messages.ts
@@ -2,8 +2,8 @@ import { defineMessages } from "react-intl";
 
 export const pluginsFilterErrorMessages = defineMessages({
   active: {
-    defaultMessage: "Active is not selected",
-    description: "plugin filters error messages active"
+    defaultMessage: "Status is not selected",
+    description: "plugin filters error messages status"
   },
   channels: {
     defaultMessage: "No channels selected",