+++ CHANGELOG.md
@@ -53,6 +53,8 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix product type translations - #1163 by @orzechdev
 - Support staff only voucher - #1174 by @orzechdev
 - Fix label names in reference attributes - #1184 by @orzechdev
+- Create channel filters in product, sales and voucher lists - #1187 by @jwm0
+- Add generic filter validation - #1187 by @jwm0
 
 # 2.11.1